refactor(nominatim): split engine, formatter, and zod wire schemas out of index.ts
